requirements.txt
streamlit==1.31.0
pandas==2.2.0
python-dotenv==1.0.0
openai>=1.25.0
pydantic>=2.0.0
requests>=2.28.0
